% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoFrame.R
\name{write.h5}
\alias{write.h5}
\title{save the cytoFrame as h5 format}
\usage{
write.h5(fr, filename)
}
\arguments{
\item{fr}{cytoFrame object}

\item{filename}{the full path of the output h5 file}
}
\description{
save the cytoFrame as h5 format
}
